moved to nested dir; added README
