--- 2023-08-08/README.md
@@ -1,4 +1,0 @@
-# Instructions for today's session
-
-
-Dataset: https://www.kaggle.com/datasets/rikdifos/credit-card-approval-prediction
