+++ 2023-08-08/credit_approval/requirements.txt
@@ -1,0 +1,8 @@
+jupyter
+matplotlib
+numpy
+pandas
+pyarrow
+scikit-learn
+sf-hamilton[visualization]
+xgboost
