--- 2023-08-08/requirements.txt
@@ -1,9 +1,0 @@
-jupyter
-matplotlib
-numpy
-optuna
-pandas
-pyarrow
-scikit-learn
-sf-hamilton[visualization]
-xgboost
